Split into binary and library (#129)
* Split into binary and library

* Split into binary and library

* Rebase to develop

* Fix tests

* Fixed builds
