+++ src/static_analysis/manifest.rs
@@ -12,8 +12,9 @@ use serde;
 use xml::reader::{EventReader, XmlEvent};
 use colored::Colorize;
 
-use {Config, Criticality, print_warning, print_vulnerability, get_code, get_string, PARSER_CONFIG};
+use {Config, print_warning, print_vulnerability, get_code, get_string, PARSER_CONFIG};
 use results::{Results, Vulnerability};
+use criticality::Criticality;
 
 pub fn manifest_analysis<S: AsRef<str>>(
     config: &Config,
